pyproject.toml
[tool.poetry]
name = "credit-risk-pd"
version = "0.1.0"
description = "Credit Risk Prediction with MLOps practices"
authors = ["Могилев Георгий"]
readme = "README.md"
packages = [{ include = "credit_risk", from = "src" }]

[tool.poetry.scripts]
credit-risk = "credit_risk.cli:app"

[tool.poetry.dependencies]
python = ">=3.10,<3.14"
python-dotenv = "^1.0.1"
omegaconf = "^2.3.0"
hydra-core = "^1.3.2"
dvc = "^3.50.2"
pre-commit = ">=4.2.0,<5.0.0"
dvc-s3 = "^3.1.0"
numpy = "<2.0"
ipykernel = "^6.29.5"
kaggle = "^1.7.4.5"
typer = "0.12.0"
pandas = "^2.2.0"
mlflow = "^2.22.1"
catboost = "^1.2.0"
optuna = "^3.3.0"
onnx = "1.14.0"
onnxmltools = "^1.11.0"
gitpython = "^3.1.0"
lightgbm = "^4.1.0"
torch = "^2.1.0"
pytorch-lightning = "^2.1.0"
torchmetrics = "^1.4.0.post0"
scikit-learn = "^1.4.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=7.4.0,<8.0.0"
pytest-cov = "^5.0.0"
black = "^24.2.0"
isort = "^5.12.0"
flake8 = "^6.1.0"
mypy = "^1.5.0"
matplotlib = "^3.8.4"

[tool.poetry.group.gpu]
optional = true

[tool.poetry.group.gpu.dependencies]
torch = "2.7.1"

[tool.poetry.group.lint]
optional = true

[tool.poetry.group.lint.dependencies]
black = "^24.2.0"
isort = "^5.13.2"
pylint = "^3.2.5"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
